updated date attribute in forms, views, and added template to confirm booking cancelation

+++ booking_system/views.py
@@ -1,22 +1,20 @@
-from django.shortcuts import render, redirect, get_object_or_404, reverse
-from django.views import generic 
 from django.contrib import messages
-from django.http import HttpResponseRedirect
-from .models import Booking
-from django.core.mail import send_mail
-from .forms import ConfirmBookingForm
 from django.contrib.auth.decorators import login_required
-from django.views.generic.edit import UpdateView 
+from django.core.mail import send_mail
+from django.http import HttpResponseRedirect
+from django.shortcuts import get_object_or_404, redirect, render, reverse
+from django.views import generic
+from django.views.generic.edit import UpdateView
 
-
-
+from .forms import ConfirmBookingForm
+from .models import Booking
 
 # Create your views here.
 
 
 def home(request):
 
-     return render(request, 'booking_system/home.html')
+    return render(request, "booking_system/home.html")
 
 
 @login_required
@@ -26,11 +24,14 @@ def booking_form(request):
         confirm_booking_form = ConfirmBookingForm(data=request.POST)
         if confirm_booking_form.is_valid():
             booking = confirm_booking_form.save(commit=False)
-            booking.guest =request.user
+            booking.guest = request.user
             booking.save()
 
-            messages.success(request, 'Your booking was submitted successfully, and will be approved shortly!')  
-    
+            messages.success(
+                request,
+                "Your booking was submitted successfully, and will be approved shortly!",
+            )
+
     confirm_booking_form = ConfirmBookingForm()
 
     return render(
@@ -40,21 +41,24 @@ def booking_form(request):
             "confirm_booking_form": confirm_booking_form,
         },
     )
-     
+
 
 @login_required
 def confirmed_booking(request):
     """
     View to display booking details
     """
-    bookings = Booking.objects.filter(approval=True, guest = request.user)
+    bookings = Booking.objects.filter(approval=True, guest=request.user)
     guest = request.user
 
-    return render(request, 'booking_system/reservation.html', 
-    {
-        'guest': guest,
-        'bookings': bookings,
-    })
+    return render(
+        request,
+        "booking_system/reservation.html",
+        {
+            "guest": guest,
+            "bookings": bookings,
+        },
+    )
 
 
 @login_required
@@ -63,16 +67,27 @@ def edit_booking(request, booking_id):
     View to edit a booking
     """
     booking = get_object_or_404(Booking, pk=booking_id)
-    form = ConfirmBookingForm(data=request.POST, instance = booking)
-    
-    if form.is_valid():
-        form.save()
-        return HttpResponseRedirect(reverse("confirmed_booking"))
+    form = ConfirmBookingForm(instance=booking)
 
-    context = {
-        "form": form
-    }
-    
+    if booking.guest != request.user:
+        messages.add_message(
+            request, messages.ERROR, "You cannot edit this booking!"
+        )
+        return redirect(reverse("homepage"))
+
+    if request.method == "POST":
+        form = ConfirmBookingForm(data=request.POST, instance=booking)
+        if form.is_valid():
+            updated_booking = form.save(commit=False)
+            updated_booking.approval = False
+            updated_booking.save()
+            messages.add_message(
+            request, messages.SUCCESS, "Your Booking was successfully Updated!"
+        )
+            return HttpResponseRedirect(reverse("confirmed_booking"))
+
+    context = {"form": form}
+
     return render(request, "booking_system/edit_booking.html", context)
 
 
@@ -83,13 +98,18 @@ def cancel_booking(request, booking_id):
     """
     booking = get_object_or_404(Booking, pk=booking_id)
 
+    if booking.guest != request.user:
+        messages.add_message(request, message.ERROR, "Booking not found!")
+        return redirect(reverse("homepage"))
+
     if request.method == "POST":
-        
-        booking.guest == request.user
         booking.delete()
-        messages.add_message(request, messages.SUCCESS, 'Your Booking was successfully deleted!')
-    else:
-        messages.add_message(request, messages.ERROR, 'Apologies, you cannot cancel this booking at the moment!')
-    
-    return HttpResponseRedirect(reverse('homepage'))
+        messages.add_message(
+            request, messages.SUCCESS, "Your Booking was successfully deleted!"
+        )
+        return redirect(reverse("homepage"))
+
+    return render(
+        request, "booking_system/cancel_confirmation.html", {"booking": booking}
+    )
